backend/requirements.txt
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
dnspython==2.6.1
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
hypothesis==6.99.9
idna==3.6
iniconfig==2.0.0
ipdata==4.0.7
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nasapy==0.2.7
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pluggy==1.4.0
Pygments==2.17.2
pymongo==4.6.2
pyperclip==1.8.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
rich==13.7.1
six==1.16.0
sortedcontainers==2.4.0
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
